medical-report-management-system/requirements.txt
Flask
Flask-SQLAlchemy
Werkzeug
pycryptodome